Registration improvements
