+++ qs-client/src/services/user/account.js
@@ -17,71 +17,13 @@ export async function signup(firstname, lastname, email, password, handle) {
       lastname: lastname,
       email: email,
       password: password,
-      handle: handle }) 
-}
-
-export async function updateProfile(id, user) {
-  return feathersClient.service("users").patch(id, user);
-}
-
-export async function forgotPassword(email) {
-  return axiosInstance.post("/authManagement", {
-    action: "sendResetPwd",
-    value: { email: email }
-  });
-}
-
-export async function updatePasswordFromProfile(
-  email,
-  oldPassword,
-  newPassword
-) {
-  return axiosInstance.post("/authManagement", {
-    action: "passwordChange",
-    value: {
-      user: {
-        email: email
-      },
-      oldPassword: oldPassword,
-      password: newPassword
-    }
-  });
-}
-
-export async function updateIdentity(password, currentEmail, updatedEmail) {
-  return axiosInstance.post("/authManagement", {
-    action: "identityChange",
-    value: {
-      password: password,
-      changes: { email: updatedEmail },
-      user: {
-        email: currentEmail
-      }
-    }
-  });
+      handle: handle 
+    }); 
 }
 
 export async function verifyAccount(token) {
   return axiosInstance.post("/authManagement", {
     action: "verifySignupLong",
     value: token
-  });
-}
-
-export async function updateUser(user, id) {
-  return feathersClient.service("users").patch(id, user);
-}
-
-export async function resendVerification(email) {
-  return axiosInstance.post("/authManagement", {
-    action: "resendVerifySignup",
-    value: { email: email }
-  });
-}
-
-export async function resetPassword(token, password) {
-  return axiosInstance.post("/authManagement", {
-    action: "resetPwdLong",
-    value: { token: token, password: password }
   });
 }
