--- qs-client/src/store/account/mutations.js
@@ -1,10 +1,0 @@
-
-function SET_USER_DATA (state, userData) {  state.user = userData
-    localStorage.setItem('user', JSON.stringify(userData))
-    axios.defaults.headers.common['Authorization'] = `Bearer ${
-      userData.token
-    }`
-};
-   
-
-
