--- qs-client/src/store/account/state.js
@@ -1,3 +1,0 @@
-export default function() {
-  return {};
-}
