+++ qs-client/src/store/modules/account/state.js
@@ -1,0 +1,6 @@
+export default function() {
+    state: {
+      user: null
+    }
+    return {};
+  }
